refactor Github's integration card to separate component
